services/drawing/pyproject.toml
[project]
name = "sona-drawing"
version = "0.1.0"
description = "Sona Drawing Command Service — FastAPI stroke DSL broadcaster"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.133.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic>=2.10.0",
    "python-dotenv>=1.0.1",
    "websockets>=15.0",
    "httpx>=0.28.0",
    "sympy>=1.14.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = []

[tool.uv]
package = false

[dependency-groups]
dev = [
    "pytest>=8.3.4",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
